REQ-403 CP-34463 prefer fqdn over ip in pool.conf
Slaves store an address to the master in pool.conf. Historically this
has been an IP address, however for REQ-403 we need to use an fqdn
instead.

Wherever pool.conf is set, we try to get an FQDN to be set instead of an
IP. We also add some code during the startup sequence that will update
any existing IPs and change them to FQDNs (this covers the upgrade
case).

HA needs a list of all host addresses, because any host could be come a
master at any point. Instead of storing all host IPs in the database, we
store all host FQDNs.

Signed-off-by: lippirk <ben.anson@citrix.com>