CA-348700: Block VDI.copy if on-boot=reset
Performing a VDI copy on a VDI that has reset-on-boot enabled may lead
to the VDI to be wiped completely, rather than reset to a predefined
state, when a VM is started. Storage migrate is blocked for VMs that
have such VDIs, and VDI.copy should have the same constraint.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>