XSI-894 handle iso8601's with no timezone
The SDK exposes our iso8601 type as an actual iso8601 value, and takes
advantage of iso8601 libraries in other languages. Therefore in order
for the SDK to work as expected, we need to align more closely with
the iso8601 specification. Internally we try our best to coerce the
iso8601 parameters into rfc3339, since rfc3339 is what ptime expects.

One particular error we have seen so far is that our date library rejects
datetime strings without a timezone qualifier (which is fixed by this change).

Signed-off-by: Ben Anson <ben.anson@citrix.com>