CA-350253: cli_operations: use `set []` when clearing if available
Currently `*-param-clear` CLI operation will retrieve the map
and individually delete entries, which is an O(N^2) operation
that can take hours, e.g. if affected by XSA-354.

We already have an API to set the map directly, so use that to set
the map to empty.

This can already be achieved through the Python API:
```
>>> import XenAPI
>>> session = XenAPI.xapi_local()
>>> session.xenapi.login_with_password("", "", "0.1", "clearme.py")
>>> vm = session.xenapi.VM.get_by_uuid("INSERT-UUID-HERE")
>>> session.xenapi.VM.set_xenstore_data(vm, {})
''
```

Signed-off-by: Edwin Török <edvin.torok@citrix.com>