REQ-403: sync certificates on xapi start
This is implemented by adding a callback at runtime to avoid laboriously
threading a callback parameter throughout the code.

The callback is guaranteed to be executed so long as the callback is mounted,
and then `Helpers.Stunnel.restart` is executed. It is also guaranteed that
the callback will be executed at most one time.

If a callback is mounted, and `Helpers.Stunnel.restart` is called
twice, then whoever reaches the callback first will execute it.

Signed-off-by: Ben Anson <ben.anson@citrix.com>