CA-351323 XSI-828 fix snapshot metadata lookup
In particular, it has been found that reverting VMs using havana
snapshots is broken (the problem was introduced in this commit:
6a431ff3368b0bd8321c99c1f852d8ad1ba4781a).

The HVM boot policy field is accessed in ocaml using HVM_boot_policy, but is
stored in the database as HVM__boot_policy (see db_actions.ml, the __ is defined
in escaping.ml). Since snapshot metadata is actually an entire database record
compressed into a single string, we must lookup the latter key in the metadata,
rather than the former.

Signed-off-by: Ben Anson <ben.anson@citrix.com>