CA-351685: improved fix for XSA-354
Introduce a byte based quota 'vm-guest-agent-xenstore-quota-bytes'.
Deprecate 'vm-guest-agent-xenstore-quota' (and convert it to bytes based
on worst case size estimates).

On startup create a table with how many bytes each character can take up
in the various JSON/XML encodings that we use, and use that to calculate
quota usage.

The default is set to the byte equivalent of the previous 128 quota, but
avoids issues where a lot of small xenstore keys/values hit the 128
quota (e.g IP addresses in OpenStack).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>