CP-33822: Use Unix.gethostname instead the hostname binary
They ultimately use the same libc function so there's no need for the
overhead of the call.

The call was used on startup and when gathering system information to
create a version number on updates. None of the functionality was using
the cached value and it was only a performance optimization.

This allows us to drop the mutex and the cached value.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>