maintenance: restrict the usage of read_localhost_info
This is to ensure that Networking_info.get_hostname is not misused
during runtime. Although effort is made in mgmt_interface to keep the
hostname value in the database sync it's better if this function is only
called when absolutely needed.

Currently it's called when the host record is created and when the
software version is generated (where it is unused)

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>