CP-36096: Generate two certificates at startup
Now two certificates are generated at startup: the additional
certificate will be used for host-to-host communications and contains
the host UUID as the issuer and subject's Common Name. Its validity
period and key type are the same as the other certificate: 3650 days and
RSA 2048 bits, respectively.

Drops gencert's dependency on mirage_crypto_rng as it doesn't generate
the private keys itself anymore.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>