CP-36100 extend update-ca-bundle to handle pool certs
update-ca-bundle.sh takes indvidicual certificates and bundles them into
a single *.pem for Stunnel to use. Extend this to handle
certificates both in

* /etc/stunnel/certs
* /etc/stunnel/certs-pool

Typically only certifcates in one of the two directories would be
updated and hence need to be re-bundled. However, the operation is so
cheap that the complexity of doing this selectively is not worth it. We
could pass an argument to this script to let it do work more
selectively.

This fixes also a bug: the directory find(1) operates on needs to exist.
The mkdir(1) call ensures it does.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>