CP-36100-2 Thread Stunnel.config value for verification
Connections may use TLS verification but the details depend on the
configuration. This patch introduces a type Stunnel.config which
describes the details to be used for verification, if at all. A value of
this type needs to be threaded through the code to create to correct
Stunnel client configuration.

Up to now client-side code for the Stunnel module was maintained in the
xen-api.git but it makes more sense to place it here as they are
conceptually related.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>