CP-36100-3 Use explicit config for TLS Stunnel verification
When Stunnel creates a conncetion as a client it can verify this
connection using certificates. We have intra pool connections and
connections opened to appliances, which use different verification
methods. This needs to be reflected beyond yes/no. This patch makes
(more) explicit the configuration to use for verifcation.

Some call sites still use no certificate checking (passing None to
SSL.make) and this will need to be adressed.

Code for configuring Stunnel client configurations was in
Helpers.StunnelClient where it is not visible to all call sites of
SSL.make. It now moved into xen-api-libs-transitional and module
Stunnel_client specifically. This open up the possibility to use it in
xe-cli, for example.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>