Enable to choose a migration network in `VM.pool_migrate`: - Add `network` key in `options` expecting a network ref to try to migrate on - If not possible, fallback on management address
Enable to choose a migration network `Host.evacuate`:
- Add `network` field to `Host.evacuate` method
- Use `network` in `VM.pool_migrate` calls when relevant

Add assert_valid_ip_configuration_on_network_for_host helper to ensure a host has a pif with a not none ip configuration in a given network.
Use this helper in host.evacaute and VM.pool_migrate when a migration network is given.

Signed-off-by: BenjiReis <benjamin.reis@vates.fr>