CA-352988: Allow localhost live migration on a disabled host
This commit allows localhost live migration on a disabled host.
The usage of this is in updating:
1. firstly the host is disabled by XAPI API users to protect this host
in following update;
2. then host.apply_updates is called. In this XAPI API call, the device
models of running HVM VMs on this host may need to be restarted. The
localhost live migration is just for this purpose.

Actually a VM.checkpoint also does a localhost migration internally,
which is allowed when the host is disabled. Localhost migration isn't
really a migration. It's just a trick to get a new domain and reload
backend services, such as the device models (QEMU).

Signed-off-by: Ming Lu <ming.lu@citrix.com>