CP-34469 on pool eject, remove host certificates
A host that leaves a pool goes trough a reboot. At boot, it will create
new host certificates if none are present. To trigger this, remove the
existing ones (private and public keys). This includes removing
(public) certificates of other hosts that were part of the pool.

We only remove certs for pool-internal communication. The user may have
updated the host certficate and installed additional CA certificates. We
keep these as this was the behaviour in CH 8.2.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>