CP-36658 remove certs of host when it is ejected
When a host leaves a pool, the remaining hosts in the pool must no
longer trust it. This is implemented by removing the leaving host's
certificates from the database and file system on all hosts in the pool.

The implementation calls Xapi_pool.eject on all hosts in the pool and
the to-be ejected host ejects itself and all others remove its
certs. The master finally removes the certificates from the database.

On the leaving host we rely on its reboot to clean up certificates.

The sequencing of this is delicate as at least leaving host and pool
master must be able to communicate during the eject call whereas all
other pool members can already remove the certificate of the leaving
host. Previously only the leaving host had to take action but now all
hosts in the pool are involved. We are careful to remove certficates
only after the eject was successful and remove the leaving host from the
pool database in the last step.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>