CP-36750: Block enabling TLS verification on pool ops
It must not be run concurrently with other pool operations as it might
disrupt the certificate distribution. It must not be run when HA is
enabled as it might fence hosts and change the master while the
operation is in progress. It is also disallowed during an RPU since
enabling it during an RPU will break the pool because the master will
not be able to contact members that have yet to upgrade.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>