CP-32863: system domain shutdown, evacuation and bacon mode support
These are storage driver domains that are host specific and get shut
down together with a host: they are not migratable because they
provide a PV storage backend for other VMs on the same host.

They are already ignored on host evacuate, and maintenance (bacon) mode the same way
that we ignore Dom0 (control domain).

On shutdown we have to proceed in 2 phases: first shut down every VM
except the driver domains, then unplug the PBDs corresponding to these
driver domain, and finally shut them down as well.
Note: the driver domain shutdown will be a hard shutdown always, but
after the PBDs are unplugged the driver domain should cope with this,
once PBD unplug is acknowledge it must've already stored all data on
persistent storage.

This ensures a clean shutdown of the system without data loss.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>