CA-353553 add API error for when NVidia GPU is misconfigured
Xapi sets up GPU types based on an XML configuration that tells it
whether a card is SR-IOV capable. When we actually use the card and
enable virtual functions and find the card is not correctly configured
for SR-IOV, we raised so far an internal error. This patch adds a new
API error to report the case more clearly.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>