CA-334758 - Allow plug of vbds for storage driver domain VMs
Cache Dom0's UUID in Xenops_server.dom0_uuid: it doesn't change
and needs to be known on each VBD plug operation.

When storage driver domains are used we need to plug VBDs into Dom0,
because the backend runs in another domain and we don't have direct access
to the VDIs.
Xenopsd only considers VMs that it manages as valid targets for VBD
plug, so Dom0 needs to be added to that condition explicitly.

The 'is managed by me' condition is not changed, because we only want to
allow this for VBD operations, and not for other operations.

Signed-off-by: bensi <ben.sims@citrix.com>
Signed-off-by: Edwin Török <edvin.torok@citrix.com>