CP-36579: Add cache in memory for available updates
This commit aims to improve the HTTP endpoint `/updates`.

Prior to this commit, HTTP endpoint `/updates` handler will always get
the available updates from all hosts in the pool. This will take quite
long duration.

With this commit, the result of available updates will be stored in a
cache in memory which is established in `pool.sync_updates`. With this
cache, HTTP endpoint `/updates` will return the result from cache
immediately.

Meanwhile, there are a few changes for management of the cache as well.

Signed-off-by: Ming Lu <ming.lu@citrix.com>