CP-36866: Generalize code for certificate distribution
Adds a Certificate provider that allows to access the filenames for a
type of certificates, adds a provider for the internal certificates used
in the current operations, adds a provider for the trusted appliance
certificates, adds a function to be able to select one of the two
providers depending on a variant type, and modifies the remote commands
to include the variant so the worker can pick the correct provider when
dealing with the filesystem.

Creates the backup directory in case in doesn't exist before trying to
copy the certificates there.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>