CA-354689 don't fail if host cert to be removed doesn't exist
Without certificate checking, there is no certificate to be removed in a
pool when a member is ejected. Previously we failed to recognise the
case where certificate checking is not enabled.

The fix tries to keep it simple rather than introducing complex code
paths based on the certificate checking beeing enabled. During pool
eject we remove certficates of the ejected host from the databse and the
file system. If certificate checking is not enabled, there is nothing to
be removed from the database and the file system. We fix here the case
for the file system by no longer failing if the file does not exist in
the first place.

The alternative design would branch on the status of certficate checking
and I believe this introduces more complexity without adding any
robustness.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>