CP-37014 verify TLS-based RPC before enabling it
Before enabling certificate checking, check that all hosts can
communicate when using it. This is implemented in a ping call that uses
certificate checking before it is enabled by default. A host pings all
pool members and all hosts do this based on code in message_forwarding.

We would like to understand how long enabling TLS verification takes
when using the O(n^2) pre-check before enabling it finally. Add some
debug output to make this easier to follow.

Add /tmp/fist_pause_after_cert_exchange fistpoint. When present, the
thread waits for it to be removed. This is an opportunity to remove
certs for testing that just have been removed.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>