CP-36178: Add basic precheck function for updates
This adds a basic call, `pool.check_update_readiness`, which returns
whether a pool is ready to be updated. It can be extended with further
checks in future, if necessary.

The call checks whether all hosts are live. Then, if the value of the
`requires_reboot` argument is `true`, it checks whether all of the hosts
can evacuated.

The result is a list of errors; an empty list implies that the pool is
ready to receive updates. Each error is encoded as a list of API error
code plus any arguments.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>