Fix SDK build
Commit 63cbb1d8 introduced a function in the datamodel that has a
return type `Set (Set String)`, which is a valid datamodel type, but
it is not yet recognised by the API bindings (SDK).

This patch just adds match cases to the code generators, such that
`make sdk` does not fail. This will need a follow-up patch to correctly
handle unmarshalling for the type.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>