CA-355571: Include accumulative updates for updates description and guidances
When applying updates on a host, there are 2 parts would be considered:
1. Latest updates, which include binary RPM packages to be installed.
The 'yum list updates' will get these RPM packages;
2. Accumulative updates. Some of these updates may be overrided by the
latest updates. These can be get from 'yum updateinfo list updates'.

E.g. pkg-a-1.0.0-r1.x86_64 has been installed on a host. There is an
update pkg-a-1.0.1-r1.x86_64 which is included in update-1. But this
update-1 has not been applied by user. Now there is another update-2
with pkg-a-1.0.2-r1.x86_64. In this case, the latest update is update-2
(pkg-a-1.0.2-r1.x86_64), the accumulative updates are update-1 and
update-2.

With this commit, the metadata (summary, description) of all
accumulative updates are included in overall update information; and
guidances of all accumulative updates are evaluated as well.
But the RPM packages to be installed are still from latest updates.

Signed-off-by: Ming Lu <ming.lu@citrix.com>