CP-37282: use xenctrl Max policy for Xen 4.7 compat code
The backward compatibility code here should've only changed
the max featureset, not the default one.
But at the time we didn't have a default/max split.

This means that all last-boot records are wrong now and contain
X86_FEATURE_IBS, which is a feature that Xen has never implemented,
but used to advertise.
In-guest we should be fine, Xen doesn't expose it anymore.
X86_FEATURE_IBS should've only ever been used to allow
for Xen 4.7 VMs to migrate.

Drop this code: xenctrl will add the necessary backward compat bits,
and only into the Max policy.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>