CP-37034: fail when featureset inconsistency is detected
Do NOT backport this to Stockholm.

To make finding bugs easier fail very early when the featuresets look
wrong: Max should always be >= default, otherwise there is probably a
bug.

On Stockholm CUx don't do this check because we might have to introduce
such a change in a security hotfix, but on master we might want to be
strict.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>