REQ-403 give cluster daemon pem information
When enabling a cluster host, or joining a cluster host to an existing
cluster, we need to get pem info from the existing cluster to give to
the enablee/joinee.

We do this by attempting to get pem info from an existing enabled
cluster host. If we are the first cluster host in a cluster then we
simple generate new pem info

Signed-off-by: Ben Anson <ben.anson@citrix.com>