CP-36097 REQ-403 add pems to cluster config
Motivation: if we add a 'pems' field to the cluster config, each cluster
member will know about a common pem file, and they can use this to
perform certificate checking

The 'blobs' field is a list, so that a cluster can trust multiple
certificates at the same time. This may be helpful in the future
if we need to implement certificate rotation inside a cluster.

We add a 'pems' parameter to the join API, so that joiners can use the
same pem file as the cluster they are trying to join.

Signed-off-by: Ben Anson <ben.anson@citrix.com>