CP-36097 REQ-403 write-pems API
Motivation for this API is the following upgrade case:

We have a stockholm cluster, which is upgraded to the latest version of
the toolstack. Before enabling TLS verification, the cluster does not
have a pem file to use, so it will use xapi's.

However in order to enable TLS verification, it will need its own pem.
'write-pems' will introduce a pem to such a cluster.

Signed-off-by: Ben Anson <ben.anson@citrix.com>