CP-36097 REQ-403 cluster must have a pem file before enabling tls verification
We enforce some constraints; for example all cluster hosts must be
enabled before trying to enable tls verification.

I don't believe this is necessary, but means we can be a little less
worried about the case where we try to re-enable a cluster node who is
missing the cluster pem (because it is less likely).

Signed-off-by: Ben Anson <ben.anson@citrix.com>