CA-355629 use hostname for CN in host cert
A self-signed certificate so puts host names and IP addess into the ANS
(alternatives names) field and the IP addess into the CN field. The ANS
field takes precedence, so the content of CN should not be relevant.
However, in the past (Stockholm release) we put the hostname into the CN
field and the change could be considered a regression. This patch adds
a hostname to the CN field if available.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>