Fix runtest and --no-buffer in xapi-idl
message-switch uses --no-buffer, while xapi-idl used buffered mode.
With --no-buffer the output is directly linked to the user's terminal
and cmdliner notices that the output is a terminal and uses a pager to
launch the help, which causes the unit tests to get stuck waiting for a
keypress.

Use `--help=plain` everywhere to avoid this.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>