XenAPI.Session: raise exception on attempted forwarding of python magic methods
Python calls magic methods for certain operations, e.g.
`not session` calls `session.__nonzero__`
and `session == foo` calls `session.__eq__`.

In the case of `__nonzero__` the failure is obvious: XAPI will return an RPC error
because it doesn't implement __nonzero__, which will be a dict and not an integer and Python raises a type error.
For a recent example of such a bug, see: https://github.com/xapi-project/sm/pull/562/commits/c52bcdf4e219126cb2b3b7a163902b85912ce4a3

`__eq__` is more problematic because the result is an error, a dict, which can be converted to an always true boolean.

Explicitly check for magic methods (starting with `__`) and reject it as not implemented.
This will give a clear stacktrace and error message, allowing to locate these bugs more easily.
Also prints the method name and arguments, which could help identify the callsite.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>