CP-34726: Add client certificate auth to stunnel config
If `pool.client_certificate_auth_enabled` is `true`, then
the incoming stunnel config is modified to verify the certificate
offered by the client. If valid, then the connection is forwarded
to a new unix socket where no further authentication is required.
The socket set up is still to do.

This authentication method is only enabled on the primary host
for the moment.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>