Stunnel.reload: wait 5s by default
The stunnel reload is initiated via systemd, which in turn sends a
signal to stunnel. This is an asynchronous call and we can't know for
sure that stunnel has reloaded its configuration when the call returns.
This can create a race condition in that we try using the new
configuration when indeed it is not yet ready. To avoid this, wait 5
seconds. There is no good mechanism to make this synchronous.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>