REQ-403 CA-356724 unix time serial number to xapi-pool-tls.pem
When stunnel/openssl sees two certificates with the same subject, it
considers them to be 'equal' in some sense. The upshot of this is that
if you tell stunnel that you would like to trust two distinct certs with
the exactly the same subject, it picks one to trust and ignores the
other. This is not desirable behaviour, so we trick stunnel into
thinking the subjects are different by setting serialNumber to the
current unix time. All that is important here is that serial numbers of
two different certs are distinct.

Signed-off-by: Ben Anson <ben.anson@citrix.com>