CA-355657 wait before serving refrehed SSL cert
After creating and distributing an internal pool certficate we reload
the Stunnel server to use it. Add waiting a few seconds before doing
that.  We might have a slow client that connects using the old cert and
has not picked up the new cert. To avoid that the connection fails,
continue using the old cert for a small time before serving the new
cert.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>