CA-349946: wait for storage driver domain to be up before plugging PBD
When plugging a PBD that has a storage driver domain attached we boot
the driver domain VM.
However we then immediately try to plug the PBD, and this can fail with
'no route to host' or 'connection refused' depending how quickly it
boots and receives an IP from the internal DHCP.

Use the already existing function in System_domains that waits until the VM is
pingable and able to respond to queries.
The VM does not need to implement XMLRPC, (the code was meant for a remote smapiv2 backend),
which is not the case here with a SMAPIv3 backend.

Require the storage driver domain VM to respond with HTTP 200 on
/HEALTHCHECK, and repeatedly query it until a timeout is hit to wait for
the boot to complete.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>