CP-37692: Introduce RBAC role for client-auth sessions
Sessions that are created on connections that are authenticated by means
of a client certificate will be assigned a new "client-cert" role.

The permissions for this new role are initially equal to those of the
pool-operator role, but adding a dedicated role will make it easier to
customise later, if desired.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>