CP-37692: Assign RBAC role to client-cert sessions
Now that there are potentially two unix sockets that xapi listens on
(the main one the new one for client-certificate auth, if enabled)
we need to be careful about the permissions that are assigned to
sessions that are created on connections to either of these sockets.

The main socket is always opened very early on in the startup sequence,
as it is used by many services in the control domain. Connections to
this socket are assumed to be trusted, as they can only come from within
dom0 itself, from a process that runs with root permissions. Sessions
created on this socket therefore do not need further authentication and
have `is_local_superuser` set to `true` to reflect their authorisation
level.

Connections to the client-cert socket are also pre-authenticated, by
stunnel verifying a client certificate, but should not be considered
root accounts. Unfortunately, it is not possible to get information from
stunnel that might distinguish various users or roles, so the best we
can do is assign a dedicated role to all sessions on this socket.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>