CA-354374: Update pool_cpuinfo and pool_features after the ejected host having been destroyed
The pool_cpuinfo and pool_feautures should be re-calculated after
the ejeced host having been destroyed in DB. Otherwise, if the
ejected host, e.g. doesn't have a common CPU feature, the pool
will not add the common CPU feature back after ejecting.

This commit fixes the issue on both pool_cpuinfo and pool_features.
Prior to this fix, the re-calculations are done by the host which
is being ejected. In this fix, they are moved to the working scope
of the master host.

Signed-off-by: minglumlu <ming.lu@citrix.com>