CP-38046: Add token in pool.sync_updates to support repository client authentication
In handling of pool.sync_updates, some HTTP requests could be sent to
the remote repositories, which could be actually hosted in a CDN, to
fetch updates.

The CDN could authenticate a request by a token. In this way, the
un-authozied users can be prevented from accessing the updates.

The rough process is:
1. a user gets the access permission as a short-lived token;
2. the user calls 'pool.sync_updates' with 2 new added (in this commit)
parameters 'token' and 'token_id';
3. the token is injected to the HTTP requests sent by YUM client as a
HTTP header 'X-Access-Token'. And the 'token_id' is injected as a HTTP
header 'Referer' to indicate the owner who is granted with the token.

Signed-off-by: Ming Lu <ming.lu@citrix.com>