CA-358583: Fix PROXY metadata to multiple HTTPS requests on a connection
When stunnel is configured with `protocol = proxy`, it will send a PROXY
header right at the beginning after establishing a connection to the
server, before passing on any decrypted data from the client. This
header contains useful TCP parameters, such as the source IP (see
04dd088). Even if this connection ends up carrying multiple HTTP(S)
requests, there will only be one PROXY header sent, as stunnel doesn't
interpret the data.

The HTTP server, however, looks for the PROXY header and HTTP header in
one step. So the first request on the connection gets annotated with
this PROXY data, and any following requests on the same connection do
not. It doesn't remember that it had already seen the PROXY header. This
is easily fixed in the receive loop by passing on a PROXY header once
one has been seen.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>