Fix handling of web-dir parameter
The web-dir parameter is not taken into account in the definition of
common_http_handlers' get_root handler, because !Xapi_globs.web_dir is
evaluated once and for all when the module loads, before the
configuration file is read.

Turn common_http_handlers into a function so that it is evaluated when
called.

Fixes #4512

Signed-off-by: Samuel Verschelde <stormi-xcp@ylix.fr>
Co-authored-by: BenjiReis <benjamin.reis@vates.fr>