CA-356506: enable verbose mode for iqn generation script
On upgrade generate-iscsi-iqn sometimes exits with code 1, but without
logging any error message.

Might be better to integrate this script into XAPI so that we get proper
stacktraces in the long run.
For now just enable verbose mode on this script so that we see exactly
how far it executed and on which line it failed in daemon.log.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>