CA-359214: Only restart stunnel if the config file has changed
This avoids unnecessarily breaking connections.

This addresses a regression introduced in d1fea38, which caused stunnel
to be restarted no matter whether a new management server was started.

The patch renames Xapi_stunnel_server.restart to .sync to reflect that
the function no longer always restarts stunnel. Also the logging is
improved.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>