CA-359835: Enable 'Updates' feature in rolling pool update
A new feature flag was added in 9e8a496f2f9f15d70f654f8af5d697a47a6f25bf.

In a rolling pool update, the first updated host (pool master) would
have this feature flag, but other pool members would not. In this
case, in the pool's perspective, the feature would be considered as
disabled. And this would block the following updates on other pool
members since this update function on pool master requires this
feature to be enabled.

The change in this commit is to set this feature enabled if other
pool members have not known this feature flag (not updated yet).

Signed-off-by: Ming Lu <ming.lu@citrix.com>