Maintenance: move xapi-storage/Makefile python rules to toplevel
Also use DESTDIR instead of PYTHON_PREFIX (which was DESTDIR+PREFIX).
Allow overriding python via PYTHON (for when we move to python3, not yet).

'make install' will now fail on regular systems because it relies on
python2, but 'make build' will continue to work.
'make install' is now the only rule that runs python ("building" python
here just copies some python files, and would rely on python2, if we
left it in 'make release' it'd break the makefile for everyone).

Note that making changes to 'xapi-storage/python/Makefile' requires
rerunning 'make build' or 'make release' for it to take effect during
'make install'.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>