CA-360577: Add RBAC checking for client cert HTTPs requests
When a HTTPs request arrived through client cert path, a 'client_cert' role
has been assigned to this request. There is no need to do authentication again
based on credentials in HTTP headers.

This commit adds RBAC checking for this kind of HTTPs requests. In details,
the `Xapi_session.login_with_password` is re-used to create a session so that
`Rbac.check` can be re-used as well to do the RBAC checking with the session.

Signed-off-by: Ming Lu <ming.lu@citrix.com>