Use stunnel proxy to access internal YUM repo
The pool members will access the internal YUM repo which is hosted on the pool
master host.

This commit aims to reuse intra-pool server certificate checking to verify
that the YUM repo is a valid one. Consequently, a stunnel client proxy will be
started on-demand to handle the TLS connection from a pool member to the pool
master, especially the server certificate checking.

Signed-off-by: Ming Lu <ming.lu@citrix.com>