CA-361220: xenopsd: introduce TASK.destroy_on_finish
There are certain tasks that are run asynchronously without anyone
waiting for the result (e.g. import_metadata_async).
Allow setting a flag on these tasks, so that they are cleaned up when
finished (either successfully or not).

Needed to avoid space leaks due to an ever growing tasks/updates list.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>